+++ services/web/server/src/simcore_service_webserver/projects/_folders_service.py
@@ -7,7 +7,7 @@ from models_library.projects import ProjectID
 from models_library.users import UserID
 
 from ..folders import _folders_repository as folders_folders_repository
-from . import _folders_repository as _folders_repository
+from . import _folders_repository
 from ._access_rights_service import get_user_project_access_rights
 from ._projects_repository_legacy import APP_PROJECT_DBAPI, ProjectDBAPI
 from .exceptions import ProjectInvalidRightsError
