+++ services/web/server/tests/integration/01/test_garbage_collection.py
@@ -41,7 +41,7 @@ from simcore_service_webserver.groups._groups_service import create_standard_gro
 from simcore_service_webserver.groups.api import add_user_in_group
 from simcore_service_webserver.login.plugin import setup_login
 from simcore_service_webserver.projects._crud_api_delete import get_scheduled_tasks
-from simcore_service_webserver.projects._groups_respository import (
+from simcore_service_webserver.projects._groups_repository import (
     update_or_insert_project_group,
 )
 from simcore_service_webserver.projects.models import ProjectDict
