+++ services/web/server/tests/unit/with_dbs/02/test_projects_comments_handlers.py
@@ -16,7 +16,7 @@ from pytest_simcore.helpers.webserver_login import LoggedUser, UserInfoDict
 from servicelib.aiohttp import status
 from simcore_service_webserver._meta import api_version_prefix
 from simcore_service_webserver.db.models import UserRole
-from simcore_service_webserver.projects._groups_respository import (
+from simcore_service_webserver.projects._groups_repository import (
     update_or_insert_project_group,
 )
 from simcore_service_webserver.projects.models import ProjectDict
